// src/payload-types.ts
export type EpisodeArray = number[];
// ...
export interface User {
  id: string;
  role: 'user' | 'admin';
  serverUrl?: string | null;
  userShows: {
    liked: number[];
    subscribed: number[];
  };
  watchedEpisodes: WatchedEpisodes;
  updatedAt: string;
  createdAt: string;
  email: string;
  username: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  _verified?: boolean | null;
  _verificationToken?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
  collection: 'users';
}
export interface WatchedEpisodes {
  [k: string]: TVShowMap;
}
/**
 * This interface was referenced by `WatchedEpisodes`'s JSON-Schema definition
 * via the `patternProperty` "^[1-9][0-9]*$".
 */
export interface TVShowMap {
  [k: string]: EpisodeArray;
}